docmind/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic>=2.9.0
httpx>=0.27.0
beautifulsoup4>=4.12.3
pymupdf>=1.24.10
nltk==3.8.1
rank-bm25==0.2.2
google-generativeai==0.7.0
python-ulid==2.7.0
aiofiles==23.2.1
python-dotenv==1.0.1
